\name{linking}

\alias{linking}

\title{Data example for IRT linking}

\description{
   Sample data for cases with two test forms. 
   The data include three common items, two items from test form A, and one item from test form B. 
   
}

\format{
A data frame in size 30 by 6.
  \describe{
    \item{c1 to c3:}{ three common items}
    \item{f1.1 to f2.1:}{ two items from test form A}    
    \item{f1.2:}{ one item from test form B}  

  }

}


\examples{

str(linking)

}


\keyword{datasets} 
