\name{person_design}

\alias{person_design}

\title{Verbal Aggression Data: Person design matrix}

\description{
   Design matrix for person covariates
}
\format{
 A data.frame for 316 cases with the following variable:

  \describe{
    \item{male}{subject's gender. male:1 and female:0}

  }

}

\references{
  De Boeck, P. and Wilson, M. (2004), Explanatory Item Response Models, Springer.
}

\examples{

str(person_design)

}


\keyword{datasets} 
