\name{science}

\alias{science}

\title{science item responses}

\description{

This data come from the British sample of the 1992 Eurobarometer
Survey (Reif and Melich, 1992) on perceptions of science and technology provided by Bartholomew et al. (2008). 
There are seven statements (such as `Science and technology are making our lives healthier, easier, and more comfortable') 
and the response categories are `strongly disagree', `disagree to
some extent', `agree to some extent', and `strongly agree'. 

Bartholomew et al. (2008) considered a between-item two-dimensional generalized partial credit model 
by using two item clusters, 1) items 1, 3, 4, and 7, and 2) items 2, 5, and 6.

}

\format{
A data frame with 392 observations with 7 items. Each item has five response categories from  `strongly disagree' to `strongly agree'. 
 
}


\references{
  Bartholomew, D. J., Steele, F., Moustaki, I., and Galbraith, J. I. (2008). 
  Analysis of Multivariate Social Science Data (Second Edition). Chapman & Hall/CRC, Boca Raton, FL.

  Reif, K. andMelich, A. (1992). Euro-barometer 38.1: Consumer protection and
  perceptions of science and technology. Technical report, INRA (Europe), Brussels.
}

\examples{

str(science)

}


\keyword{datasets}
