\name{spelling}

\alias{spelling}

\title{spelling item responses}

\description{
   This data come from Thissen et al. (1993, p.71) that examined performance of under-
   graduates on four spelling items `infidelity', `panoramic', `succumb' and
   `girder'. Each spelling item was scored either as correct or incorrect.
   The sample includes 285 male and 374 female undergraduate students from the University of Kansas. 

}

\format{
A data frame with 659 observations on the four spelling items. 
There is one more variable, \code{male} that takes value 1 for male and 0 for female.
 
}

\references{
Thissen, D., L. Steinberg, and H. Wainer. 1993. Detection of differential item functioning
using the parameters of item response models. In Differential Item Functioning, ed.
P. Holland and H. Wainer, 67-114. Hillsdale, NJ: Lawrence Erlbaum.
}

\examples{

str(spelling)

}


\keyword{datasets}

