\name{verb2}

\alias{verb2}

\title{Verbal Aggression Data: binary responses}

\description{
   Data frame in wide form for the verbal aggression data. 
}
\format{
 A data frame for 316 cases with the following variables:

  \describe{
    \item{y1-y24}{binary item responses to item 1 to item 24. 
    The order of items are the same as that in VerbAgg (long form). No:0, Perhaps and Yes:1 } 
    
  }

}

\references{
  De Boeck, P. and Wilson, M. (2004), Explanatory Item Response Models, Springer.
}

\examples{

str(verb2)

}


\keyword{datasets} 
