\name{verb3}

\alias{verb3}

\title{Verbal Aggression Data: polytomous responses}

\description{
   Data frame in wide form for the verbal aggression data. 
}
\format{
A data frame for 316 cases with the following variables. 

  \describe{
    \item{y1-y24}{original polytomous item responses to item 1 to item 24. 
    The order of items are the same as that in VerbAgg (long form). No:0, Perhaps:1, and Yes:2  } 
    
  }

}

\references{
  De Boeck, P. and Wilson, M. (2004), Explanatory Item Response Models, Springer.
}

\examples{

str(verb3)

}


\keyword{datasets} 
